flask-project/openstack-create-project-ad-net-ok/requirements.txt
openstacksdk
paramiko